nir/lower_int64: explicitly cast integers to doubles
This eliminates a few of these warnings on Clang:

warning: implicit conversion from 'unsigned long' to 'double' changes value
from 18446744073709551615 to 18446744073709551616
[-Wimplicit-const-int-float-conversion]

warning: implicit conversion from 'long' to 'double' changes value from
9223372036854775807 to 9223372036854775808
[-Wimplicit-const-int-float-conversion]

+++ src/compiler/nir/nir_lower_int64.c
@@ -760,13 +760,13 @@ lower_f2(nir_builder *b, nir_ssa_def *x, bool dst_is_signed)
    if (dst_is_signed)
       x_sign = nir_fsign(b, x);
    else
-      x = nir_fmin(b, x, nir_imm_floatN_t(b, UINT64_MAX, x->bit_size));
+      x = nir_fmin(b, x, nir_imm_floatN_t(b, (double)UINT64_MAX, x->bit_size));
 
    x = nir_ftrunc(b, x);
 
    if (dst_is_signed) {
-      x = nir_fmin(b, x, nir_imm_floatN_t(b, INT64_MAX, x->bit_size));
-      x = nir_fmax(b, x, nir_imm_floatN_t(b, INT64_MIN, x->bit_size));
+      x = nir_fmin(b, x, nir_imm_floatN_t(b, (double)INT64_MAX, x->bit_size));
+      x = nir_fmax(b, x, nir_imm_floatN_t(b, (double)INT64_MIN, x->bit_size));
       x = nir_fabs(b, x);
    }
